Separate the validate function and do not use from node
