+++ src/index.ts
@@ -3,3 +3,4 @@
 
 export * from './builder';
 export * from './plugin';
+export * from './validate';
