+++ test/src/builder.spec.ts
@@ -4,53 +4,13 @@
 import expect = require('expect.js');
 
 import {
-  buildExtension, validateEntry
+  buildExtension
 } from '../../lib';
 
 
 describe('builder', () => {
 
-  describe('validateEntry()', () => {
-
-    it('should pass for a valid plugin array', () => {
-      validateEntry([{
-        id: 'foo',
-        activate: () => { /* no-op */ }
-      }, {
-        id: 'bar',
-        activate: () => { /* no-op */ }
-      }]);
-    });
-
-    it('should pass for a valid plugin', () => {
-      validateEntry({
-        id: 'foo',
-        activate: () => { /* no-op */ }
-      });
-    });
-
-    it('should pass for an ES6 default', () => {
-      validateEntry({
-        __esModule: true,
-        default: {
-          id: 'foo',
-          activate: () => { /* no-op */ }
-        }
-      });
-    });
-
-    it('should fail if it is an empty array', () => {
-      expect(() => { validateEntry([]); }).to.throwError();
-    });
-
-    it('should fail if a plugin is missing an id', () => {
-      let activate: () => { /* no-op */ };
-      expect(() => { validateEntry({ activate }); }).to.throwError();
-    });
-
-    it('should fail if a plugin is missing an activate function', () => {
-      expect(() => { validateEntry({ id: 'foo' }); }).to.throwError();
-    });
+  it('should pass', () => {
 
   });
